[ANDROSDK-1056] Give meaningful names to database objects in access folder
